social-media-agent/requirements.txt
youtube_transcript_api
google-genai
